Moving pyomo_util to the solvers subdir
This logic is only used to support the implementation of solvers using
Pyomo.
