+++ pao/lbp/__init__.py
@@ -2,7 +2,6 @@ from .repn import LinearMultilevelProblem, QuadraticMultilevelProblem
 from . import convert_repn
 from .solver import SolverFactory
 #from . import pyomo_solvers
-from . import pyomo_util
 from . import examples
 from . import solvers
 from . import soln_manager
