+++ pao/lbp/solvers/ld.py
@@ -11,7 +11,7 @@ import pyomo.opt
 from ..solver import SolverFactory, LinearMultilevelSolverBase, LinearMultilevelResults
 from ..repn import LinearMultilevelProblem
 from ..convert_repn import convert_LinearMultilevelProblem_to_standard_form
-from .. import pyomo_util
+from . import pyomo_util
 
 nan = float('nan')
